pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "taxbynav-backend"
version = "0.1.0"
description = "Backend for TaxByNav - FastAPI, JWT auth, appointments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "python-multipart>=0.0.17",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "psycopg2-binary>=2.9.0",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.uv.scripts]
dev = "uvicorn app.main:app --reload"
lint = "ruff check app tests"
format = "ruff format app tests"
migrate = "alembic upgrade head"
migrate-gen = "alembic revision --autogenerate -m"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
